requirements.txt
fastapi
uvicorn
langchain
langchain-groq
sentence-transformers
pymupdf
chromadb
python-dotenv
